Move `driver` package into `server` package (#2926)
