+++ e2e/opacus/driver.py
@@ -1,6 +1,6 @@
 import flwr as fl
 
-hist = fl.driver.start_driver(
+hist = fl.server.driver.start_driver(
     server_address="0.0.0.0:9091",
     config=fl.server.ServerConfig(num_rounds=3),
 )
