+++ examples/quickstart-cpp/driver.py
@@ -3,7 +3,7 @@ from fedavg_cpp import FedAvgCpp
 
 # Start Flower server for three rounds of federated learning
 if __name__ == "__main__":
-    fl.driver.start_driver(
+    fl.server.driver.start_driver(
         server_address="0.0.0.0:9091",
         config=fl.server.ServerConfig(num_rounds=3),
         strategy=FedAvgCpp(),
