+++ examples/secaggplus-mt/driver.py
@@ -6,7 +6,7 @@ import numpy as np
 from workflows import get_workflow_factory
 
 from flwr.common import Metrics, ndarrays_to_parameters
-from flwr.driver import GrpcDriver
+from flwr.server.driver import GrpcDriver
 from flwr.proto import driver_pb2, node_pb2, task_pb2
 from flwr.server import History
