+++ src/py/flwr/__init__.py
@@ -17,12 +17,11 @@
 
 from flwr.common.version import package_version as _package_version
 
-from . import client, common, driver, server, simulation
+from . import client, common, server, simulation
 
 __all__ = [
     "client",
     "common",
-    "driver",
     "server",
     "simulation",
 ]
