+++ src/py/flwr/server/driver/app_test.py
@@ -20,13 +20,13 @@ import time
 import unittest
 from unittest.mock import MagicMock
 
-from flwr.driver.app import update_client_manager
 from flwr.proto.driver_pb2 import (  # pylint: disable=E0611
     CreateRunResponse,
     GetNodesResponse,
 )
 from flwr.proto.node_pb2 import Node  # pylint: disable=E0611
 from flwr.server.client_manager import SimpleClientManager
+from flwr.server.driver.app import update_client_manager
 
 
 class TestClientManagerWithDriver(unittest.TestCase):
