+++ src/py/flwr/server/driver/driver.py
@@ -17,7 +17,6 @@
 
 from typing import Iterable, List, Optional, Tuple
 
-from flwr.driver.grpc_driver import DEFAULT_SERVER_ADDRESS_DRIVER, GrpcDriver
 from flwr.proto.driver_pb2 import (  # pylint: disable=E0611
     CreateRunRequest,
     GetNodesRequest,
@@ -26,6 +25,7 @@ from flwr.proto.driver_pb2 import (  # pylint: disable=E0611
 )
 from flwr.proto.node_pb2 import Node  # pylint: disable=E0611
 from flwr.proto.task_pb2 import TaskIns, TaskRes  # pylint: disable=E0611
+from flwr.server.driver.grpc_driver import DEFAULT_SERVER_ADDRESS_DRIVER, GrpcDriver
 
 
 class Driver:
